Update JavaScript for Traffic Count attachments project
This version of the Data Tracker's javascript page contains the code used to extract the Attachment Type category names from that column and replace the filenames of the attachments (in the Attachments column) with those extracted names as requested.  I made a different copy of the code for each page which is not ideal, but I'm not immediately sure how to improve on that because I'm not sure if I can declare multiple cases for $(document).on to occur in.  If that does work it would make the code a lot cleaner, we could just include every page we need the code to run on in one copy of the code instead of making a separate case for each page.  This version does work well as it is, though.